# [perf] optimize ValueAtPercentile(s): introduce nextCountAtIdx() and reduce by 25.6% the on-cpu usage