This PR optimizes ValueAtPercentile(s) by introducing nextCountAtIdx() and only updating the iterator highestEquivalentValue when required in order to reduce wasteful cpu usage. If we look at ValueAtPercentile numbers, It reduced by 25.6% the on-cpu usage.

master branch benchmark percentiles calculation:
```
% go test -bench="BenchmarkHistogramValueAtPercentile" -benchtime=10s -cpuprofile profile.out
goos: darwin
goarch: arm64
pkg: github.com/HdrHistogram/hdrhistogram-go
BenchmarkHistogramValueAtPercentile-8                        	  264561	     42905 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramValueAtPercentileGivenPercentileSlice-8    	   58002	    214438 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramValueAtPercentilesGivenPercentileSlice-8   	  173206	     69352 ns/op	     248 B/op	       4 allocs/op
```

cpu.optimize.nextCountAtIdx branch benchmark percentiles calculation:
```
% go test -bench="BenchmarkHistogramValueAtPercentile" -benchtime=10s -cpuprofile profile.out
goos: darwin
goarch: arm64
pkg: github.com/HdrHistogram/hdrhistogram-go
BenchmarkHistogramValueAtPercentile-8                        	  369895	     31900 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramValueAtPercentileGivenPercentileSlice-8    	   83106	    147954 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramValueAtPercentilesGivenPercentileSlice-8   	  232894	     52160 ns/op	     248 B/op	       4 allocs/op
```